evals: keep project audit in eval CLI
